Regulierungen

Dieser Datensatz beinhaltet Informationen des VSA zur Regulierung von Mikroverunreinigungen.

Verwendung

regulierungen_mvwizr

Format

regulierungen_mvwizr

Tabelle mit 4 Variablen und knapp 3000 Einträgen. Im Folgenden sind die Variablen mit ihren Datentypen beschrieben:

ID_Substanz

VSA-ID der Substanz (integer)

Name_reg

Bezeichnung, die vom VSA verwendet wird (character)

Informationen Recht

Zeichenkette mit codierten Regulierungsinformationen (character)

GSCHV

Gibt an, ob und wie die Substanz in der GSCHV reguliert ist. 1 = Spezifischer Wert in Anh. 2; 2 = allgemeiner Grenzwert von 0.1 µg/l (integer)

Quelle

VSA